moved release logic
